% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_add_failed_plates.R
\name{add_failed_plates}
\alias{add_failed_plates}
\title{Add Failed plates}
\usage{
add_failed_plates(results, image_dir)
}
\arguments{
\item{results}{output of number_plate_match()}

\item{image_dir}{directory of image files}
}
\description{
takes the results of the numberplate validation and allows teh addition of plates that were not detected by ALPR
}
\author{
W S Drysdale
}
