% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_aplr_remove_duplicates.R
\name{aplr_remove_duplicates}
\alias{aplr_remove_duplicates}
\title{ALPR Remove duplicates}
\usage{
aplr_remove_duplicates(results)
}
\arguments{
\item{results}{output of run_openALPR()}
}
\description{
Where the ALRP software has returned multiple results, keeps only the highest confidence one
}
\author{
W S Drysdale
}
