% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_openALPR_request.R
\name{openALPR_request}
\alias{openALPR_request}
\title{API Request function}
\usage{
openALPR_request(secret_key, country = "eu", image)
}
\arguments{
\item{secret_key}{secret key of users account}

\item{country}{country code of plates, default eu}

\item{image}{file path of image}

\item{prewarp}{prewarp coordiantes, defaults to no prewarp}
}
\description{
Handles calls to the openALPR cloud API
}
\author{
W. S. Drysdale
}
