% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_parse_openALPR_responses.R
\name{parse_openALPR_responses}
\alias{parse_openALPR_responses}
\title{Parse open APLR response}
\usage{
parse_openALPR_responses(responses, image_files)
}
\arguments{
\item{responses}{list containing the output of openAPLR_request}

\item{image_files}{vector of image file names that was supplie to openAPLR_request}
}
\description{
handles the response from the openAPLR cloudAPI response and returns useable results
}
\author{
W. S. Drysdale
}
