% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_process_image.R
\name{process_image}
\alias{process_image}
\title{Process Image}
\usage{
process_image(image_file, output_folder, vcrop = NULL, brightness = 0,
  contrast = 1)
}
\arguments{
\item{image_file}{path to one image}

\item{output_folder}{where should these processed images be saved?}

\item{vcrop}{vertical crop range to "letterbox" image - default 30:220}

\item{brightness}{value to add or subract to adjust image brightness via EBImage}

\item{contrast}{value to multiply by to adjust image contrat via EBImage}
}
\description{
Wrapper function for processing images - allows for adjustment of processing options from \code{run_openALPR()}
}
\author{
W. S. Drysdale
}
