% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_run_openALPR.R
\name{run_openALPR}
\alias{run_openALPR}
\title{Run openAPLR}
\usage{
run_openALPR(secret_key, country = "eu", image_dir)
}
\arguments{
\item{secret_key}{openALPR API secret key}

\item{country}{openALPR plate country code, default "eu"}

\item{image_dir}{directory where image files are stored}
}
\description{
Runs FEAT images though the openALPR system and outputs a respose ready to be parsed through verification steps
}
\author{
W S Drysdale
}
